+++ src/app/components/modals/file-chunks.js
@@ -23,10 +23,11 @@ const {
   run,
   inject: { service },
   assert,
+  RSVP: { Promise },
 } = Ember;
 
 const SLOW_POLLING_TIME = 10 * 1000;
-const FAST_POLLING_TIME = 2 * 1000;
+const FAST_POLLING_TIME = 4 * 1000;
 
 export default Component.extend(PromiseLoadingMixin, {
   classNames: ['file-chunks', 'file-chunks-modal'],
@@ -211,36 +212,63 @@ export default Component.extend(PromiseLoadingMixin, {
    * @type {SpaceTransferList}
    */
   currentTransferList: computed.reads('space.currentTransferList'),
+  
+  scheduledTransferList: computed.reads('space.scheduledTransferList'),
 
   /**
    * Array of current transfers
    * @type {Array<Transfer>} with isLoaded Ember property
    */
-  currentTransfers: computed.reads('currentTransferList.list.content'),
+  workingTransfers: computed(
+    'workingTransfersDataLoaded',
+    'scheduledTransferList.list.content.[]',
+    'currentTransferList.list.content.[]',
+    'lastTransfer',
+    function () {
+      if (this.get('workingTransfersDataLoaded')) {
+        const scheduled = this.get('scheduledTransferList.list.content');
+        const current = this.get('currentTransferList.list.content');
+        const lastTransfer = this.get('lastTransfer');
+        if (scheduled && current) {
+          this.set('lastTransfer', null);
+          return (lastTransfer ? [lastTransfer] : []).concat(scheduled.toArray(), current.toArray());
+        }
+      }
+    }
+  ),
 
   /**
    * True if each transfer is ready to be inserted into table, _but_ without
    * async dynamic data like transferred bytes or status
    * @type {Ember.ComputedProperty<boolean>}
    */
-  currentTransfersDataLoaded: computed(
+  workingTransfersDataLoaded: computed(
+    'scheduledTransferList.isLoaded',
     'currentTransferList.isLoaded',
-    'currentTransfers.@each.isLoaded',
+    'scheduledTransferList.list.content.isLoaded',
+    'currentTransferList.list.content.isLoaded',
     function getCurrentTransfersDataLoaded() {
       return this.get('currentTransferList.isLoaded') === true &&
-        this.get('currentTransfers').every(t => get(t, 'isLoaded') === true);
+        this.get('scheduledTransferList.isLoaded') === true &&
+        this.get('currentTransferList.list.content.isLoaded') === true &&
+        this.get('scheduledTransferList.list.content.isLoaded') === true;
     }
   ),
 
   /**
    * @type {Ember.ComputedProperty<Array<Transfer>>}
    */
-  fileTransfers: computed('currentTransfersDataLoaded', 'currentTransfers.[]', function () {
-    const currentTransfersDataLoaded = this.get('currentTransfersDataLoaded');
-    const currentTransfers = this.get('currentTransfers');
+  fileTransfers: computed('workingTransfersDataLoaded', 'workingTransfers.[]', function () {
+    const {
+      workingTransfersDataLoaded,
+      workingTransfers,
+    } = this.getProperties(
+      'workingTransfersDataLoaded',
+      'workingTransfers'
+    );
     const fileId = this.get('file.id');
-    if (currentTransfersDataLoaded && currentTransfers) {
-      return currentTransfers.filter(t => t.belongsTo('file').id() === fileId);
+    if (workingTransfersDataLoaded && workingTransfers) {
+      return workingTransfers.filter(t => t.belongsTo('file').id() === fileId);
     }
   }),
   
@@ -305,6 +333,7 @@ export default Component.extend(PromiseLoadingMixin, {
       store,
       space,
       isEnabled: true,
+      scheduledEnabled: true,
       currentEnabled: true,
       completedEnabled: false,
     }));
@@ -343,7 +372,7 @@ export default Component.extend(PromiseLoadingMixin, {
    * @returns {Promise<Array<FileDistribution>>}
    */
   fetchDistribution() {
-    console.log('component:modals/file-chunks: fetchDistribution');
+    console.debug('component:modals/file-chunks: fetchDistribution');
     const fileId = this.get('file.id');
     this.get('store').query('file-distribution', { file: fileId }).then(
       (fbs) => {
@@ -434,6 +463,21 @@ export default Component.extend(PromiseLoadingMixin, {
     }
   },
   
+  forceUpdateTransfers(transfersUpdater) {
+    return Promise.all([
+        transfersUpdater.fetchScheduled(),
+        transfersUpdater.fetchCurrent(true)
+      ])
+      .catch(error => {
+        // TODO: i18n
+        this.set('chunksModalError', 'Loading transfers data failed');
+        throw error;
+      })
+      .finally(() => {
+        this.set('transfersLoading', false);
+      });
+  },
+  
   actions: {
     /**
      * File chunks modal component is placed all the time,
@@ -445,15 +489,7 @@ export default Component.extend(PromiseLoadingMixin, {
       
       const transfersUpdater = this._initTransfersUpdater();
       this.set('transfersLoading', true);
-      transfersUpdater.fetchCurrent(true)
-        .catch(error => {
-          // TODO: i18n
-          this.set('chunksModalError', 'Loading transfers data failed');
-          throw error;
-        })
-        .finally(() => {
-          this.set('transfersLoading', false);
-        });
+      this.forceUpdateTransfers(transfersUpdater);
 
       this._initDistributionUpdater();
       this.fetchDistribution();
@@ -524,9 +560,8 @@ export default Component.extend(PromiseLoadingMixin, {
           this.observeTransfersCount();
           throw error;
         })
-        .then(( /* transfer */ ) => {
-          return transfersUpdater.fetchCurrent();
-        })
+        .then(() => transfersUpdater.fetchScheduled())
+        .then(() => transfersUpdater.fetchCurrent())
         .then(() => {
           return this._fastTransfersUpdater();
         })
@@ -556,12 +591,11 @@ export default Component.extend(PromiseLoadingMixin, {
           this.observeTransfersCount();
           throw error;
         })
-        .then(( /* transfer */ ) => {
-          return transfersUpdater.fetchCurrent();
+        .then(transfer => {
+          this.set('lastTransfer', transfer);
+          return this.forceUpdateTransfers(transfersUpdater);
         })
-        .then(() => {
-          return this._fastTransfersUpdater();
-        });
+        .then(() => this._fastTransfersUpdater());
     },
   },