+++ src/app/locales/en/translations.js
@@ -502,13 +502,18 @@ export default {
       publish: 'Publish',
     },
     transfers: {
-      transfersFor: 'Transfer for',
       providersMapOfDist: 'Active transfers map',
+      scheduledTransfers: 'Scheduled transfers',
       activeTransfers: 'Active transfers',
       completedTransfers: 'History of transfers',
       onTheFlyTransfers: 'On-the-fly transfers',
       noActiveTransfers: 'There are no active transfers',
       noCompletedTransfers: 'There are no past transfers',
+      noTransfers: {
+        scheduled: 'There are no scheduled transfers',
+        current: 'There are no active transfers',
+        completed: 'There are no past transfers',
+      },
       initializingTransfers: 'Initializing transfers...',
       notSupported: 'Cannot list transfers of selected space because it is not ' +
         'supported by current provider',
@@ -519,6 +524,7 @@ export default {
         path: 'File/directory',
         userName: 'Username',
         destination: 'Destination',
+        scheduledAt: 'Scheduled at',
         startedAt: 'Started at',
         finishedAt: 'Finished at',
         totalBytes: 'Transferred',
@@ -566,6 +572,10 @@ export default {
         onTheFly: 'On-the-fly',
         all: 'All',
         jobs: 'Jobs',
+      },
+      transfersOverview: {
+        hide: 'Hide transfers overview',
+        show: 'Show transfers overview',
       },
     },
     rootDirSettingsDrop: {